Regular update
 * revised vendor projects
 * trying out BiF
 * updated submodules

+++ build/bro.sh
@@ -3,7 +3,9 @@
 set -ex
 
 # download Bro
-wget https://www.zeek.org/downloads/bro-2.6.1.tar.gz -O bro-2.6.1.tar.gz
+if [ ! -f bro-2.6.1.tar.gz ] ; then
+    wget https://www.zeek.org/downloads/bro-2.6.1.tar.gz -O bro-2.6.1.tar.gz
+fi
 tar -xzf bro-2.6.1.tar.gz
 cd bro-2.6.1
 
@@ -13,7 +15,11 @@ pipenv run ./configure \
     --prefix=$(pipenv --venv)/bro \
     --localstatedir=$(pipenv --venv)/var \
     --conf-files-dir=$(pipenv --venv)/etc \
-    --with-openssl=$(brew --prefix openssl)
+    --with-openssl=$(brew --prefix openssl) \
+    --with-bison=$(brew --prefix bison)/bin/bison \
+    --with-python=$(pipenv --py) \
+    --with-geoip=$(brew --prefix geoip) \
+    --with-swig=$(brew --prefix swig)/bin/swig
 pipenv run make
 pipenv run make install
